Move testing util functions from userActionsTestData to utils/reduxTestingFunctions
