+++ src/redux/actions/userActions1.test.js
@@ -1,11 +1,11 @@
 import * as actions from "./userActions";
 import * as types from "../types/userTypes";
 import { APIMock } from "../../utils/APIMock";
+import { testData as data } from "./userActionsTestData";
 import {
-  testData as data,
   compareExpectedCalls,
   spreadCalls,
-} from "./userActionsTestData";
+} from "../../utils/reduxTestingFunctions";
 
 const getState = jest.fn();
